voice_gateway/requirements.txt
# Voice Gateway specific requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
httpx==0.27.0
redis==5.0.1

# Audio & STT/TTS
webrtcvad==2.0.10
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
groq==0.36.0
openai==1.12.0
numpy==1.26.3
scipy==1.12.0
twilio==8.11.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
python-dateutil==2.8.2
python-multipart==0.0.6
